refactor: :recycle: asset/icon 에서 import 하도록 수정
