+++ src/assets/icon.ts
@@ -1,3 +1,5 @@
 export { ReactComponent as HistoryIcon } from './history.svg';
 export { ReactComponent as SearchIcon } from './search.svg';
 export { ReactComponent as FlightTakeOffIcon } from './flight_takeoff.svg';
+export { ReactComponent as BookmarkIcon } from './bookmark.svg';
+export { ReactComponent as BookmarkFillIcon } from './bookmark_fill.svg';
